backend/pyproject.toml
[project]
name = "telegram-agents-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "PyJWT>=2.10.0",
    "bcrypt>=4.2.0",
    "aiogram>=3.14.0",
    "anthropic>=0.40.0",
    "openai>=1.57.0",
    "arq>=0.26.1",
    "redis>=5.2.0",
    "python-multipart>=0.0.12",
    "httpx>=0.28.0",
]
